refactor tweet video
moved tweet video funciton out main index.js and into TweetWithVideo module
