+++ TODO.md
@@ -1,6 +1,6 @@
 # TODOS
 
-- [ ] move `tweetClipHelper` to module `TweetWithVideo` to promisify module. 
+- [X] move `tweetClipHelper` to module `TweetWithVideo` to promisify module. 
 - [X] index.js promisify the overall module
-
-- [ ] update PR
+- [X] update PR
+- [ ] clean up tmp files, eg set tmp directory?
