+++ lib/TweetWithVideo/example-usage-callback.js
@@ -2,15 +2,13 @@
  * Example usage, 
  * Instantiates a TweetWithVideo
  */
-
-
 // load credentials
 const dotenv = require('dotenv');
 const loadEnv = dotenv.config();
 // error handlign credentials 
 if (loadEnv.error) throw loadEnv.error;
 
-const TweetWithVideo = require('./index.js');
+const TweetWithVideo = require('./TweetWithVideo.js');
 const path = require('path');
 
 // auth 
@@ -28,3 +26,4 @@ videoTweet = new TweetWithVideo({
 }, (error, response)=>{
     console.log('example usage response: ',response);
 });
+
