+++ lib/burn-captions/index.js
@@ -5,7 +5,6 @@
 const spawn = require('child_process').spawn;
 
 function burnCaptions(opts){
-    // console.log(options)
     // return a promise 
     return new Promise((resolve, reject)=>{
         // TODO: Font type, and font size could be passed in as optional params
